+++ rendererrecyclerviewadapter/src/main/java/com/github/vivchar/rendererrecyclerviewadapter/BinderAdapter.java
@@ -1,8 +1,6 @@
-package com.github.vivchar.rendererrecyclerviewadapter.binder;
+package com.github.vivchar.rendererrecyclerviewadapter;
 
 import androidx.annotation.NonNull;
-
-import com.github.vivchar.rendererrecyclerviewadapter.ViewRenderer;
 
 import java.util.List;
