+++ rendererrecyclerviewadapter/src/main/java/com/github/vivchar/rendererrecyclerviewadapter/OnClickListener.java
@@ -1,0 +1,7 @@
+package com.github.vivchar.rendererrecyclerviewadapter;
+
+public interface OnClickListener {
+    void onClick();
+}
+
+
