+++ rendererrecyclerviewadapter/src/main/java/com/github/vivchar/rendererrecyclerviewadapter/WrongViewFinderException.java
@@ -1,4 +1,4 @@
-package com.github.vivchar.rendererrecyclerviewadapter.binder;
+package com.github.vivchar.rendererrecyclerviewadapter;
 
 /**
  * Created by Vivchar Vitaly on 11.07.19.
