--- rendererrecyclerviewadapter/src/main/java/com/github/vivchar/rendererrecyclerviewadapter/binder/OnClickListener.java
@@ -1,7 +1,0 @@
-package com.github.vivchar.rendererrecyclerviewadapter.binder;
-
-public interface OnClickListener {
-    void onClick();
-}
-
-
